对database进行细化

+++ ASF_Plus/app/basic.py
@@ -1,5 +1,5 @@
 # 填写ASF密码在这
-pw = 'pw'
+pw = 'dalizm8881018'
 
 class asfbot:
 	def __init__(self,name):
@@ -31,25 +31,56 @@ class asfbot:
 		
 	def delBot(self):
 		path = os.path.join(getConfigPath(),f'{self.name}.cfg')
-		os.remove(path)
+		if os.path.isfile(path):
+			os.remove(path)
 		self.update()
+		
+	def createBot(self,ac,pw,ow):
+		if self.exist:
+			return False
+		data = {'BotConfig':{'SteamLogin':ac,'SteamPassword':pw}}
+		resp = asf_re('Bot',self.name,data,True)
+	
+	def loginBot(self,tfa=''):
+		if not self.exist:
+			pass # change it after debuging
+		if tfa:
+			data = {'Type':'TwoFactorAuthentication','Value':tfa}
+			asf_re('Bot',self.name+'/Input',data,True)
+		else:
+			asf_re('Bot',self.name+'/Start',post=True)
+	#	if resp[self.name]:
+		#	return True
+		
+		
+		
 		
 
 
 import requests, json, os
 
 
-def asf_re(par,bot='',req=[]):
+def asf_re(par,bot='',req=[],post=False):
 	if bot:
 		url = f'https://asf.novanoir.cn/Api/{par}/{bot}?password={pw}'
 	else:
 		url = f'https://asf.novanoir.cn/Api/{par}?password={pw}'
-	try:
-		with requests.get(url, timeout=20) as resp:
-			res = resp.json()
-			return res['Result']
-	except:
-		return False
+	if not post:
+		try:
+			with requests.get(url, timeout=20) as resp:
+				res = resp.json()
+				print(res)
+				return res['Result']
+		except:
+			return False
+	else:
+		try:
+			with requests.post(url, timeout=20,json=req) as resp:
+				res = resp.json()
+				print(res)
+				return res['Result']
+		except:
+			return False
 		
 def asf_addBot(bot,owner):
 	path = os.path.join(getConfigPath(),f'{bot}.cfg')
